Make effectEngine isActive function a pureComputed
This is to prevent the returned function being forever held in memory.
A computed will still be referenced in the subscriber list for the item,
so the garbage collection never knows it can be cleaned up.
A pureComputed will unsubscribe from everything when there is no longer
anything subscribed to itself.