feat(typescript): initial work to support TypeScript modules
This is the initial split of our code from `./src/scripts` (old code
that assumes classes and types are global variables) to `./src/modules`
(TypeScript code using import/export). The main goal of this is to allow
us to gain full typechecking and bundling benefits, and to reduce the
circular dependencies that have only been allowed to exist due to the
order in which code runs.