Add a link to the discord in the readme (#882)
Add a link to the discord in the readme and update link to game

If you arrive at the github repo, it would be good to immediately be
able to find the development community to discuss ideas and how to go
about implementing them. This commit adds the invite link currently
listed on the subreddit.