feat(typescript): initial work to support TypeScript modules (#871)
* feat(typescript): initial work to support TypeScript modules

This is the initial split of our code from `./src/scripts` (old code
that assumes classes and types are global variables) to `./src/modules`
(TypeScript code using import/export). The main goal of this is to allow
us to gain full typechecking and bundling benefits, and to reduce the
circular dependencies that have only been allowed to exist due to the
order in which code runs.

* feat(typescript): move Statistics into TS module

* chore(ci): ensure the code is tested with modules+scripts

* fix(typescript): fix paths so gulp works in Windows

* chore(typescript): clean declarations before generating new ones

If a module is moved or deleted, the declaration will cause scripts to
crash due to duplicate declarations. By cleaning the declarations
directory prior to generating them, we will ensure there are only the
most up-to-date declarations available