Initial SVG implementation for Unova
Notes:
Not entirely sure what town exists where Black City/White Forest is. I've left it as "???" for now.
There doesn't seem to be a dock in Unova right now. Not sure if that just wasn't implemented yet, so I commented out the dock for now.
Castelia city has a dungeon, and what seems to be some type of machine/gate thing near it. I'm not sure what both of these are referring to, but I added elements for both.
There are comments for Desert Resort and Undella Bay to become routes. I've still added elements for both dungeons for now.