feat(statistics): track routeKills as a structured object
This will store the statistics for `routeKills` as an object keyed by a
region, and an object keyed by the route, with the value being the number of
kills (`Record<string, Record<number, number>>`).

```
{
  kanto: {
    1: 15,
    2: 10,
  },
  johto: {
    26: 20,
    27: 13,
  },
}
```

These will be automatically turned into an observable when loaded from
the save, and automatically converted back into a JSON object when
generating a save. The proxy for `routeKills` will allow accessing
statistics in a region by the region enum's string or number value.