Add totems as post-league bosses
Totems get an "aura boost". In the main series this results in 1 or 2 stages of increase into specific stats. A stage is 50%. So that's 50% or 100% boost. For the stats of the totems in clicker I've turned every 1 into +10, ever 2 into +20. Just added to the stats. No percentages. Can't go around doubling stats, considering these totems will be catchable.