fixes(farm): Berry fix round 4 (#1066)
* Fix coin image for Oak Items Display

* Increment Berry harvest statistic on withers as well

* Remove Combee from Shop

Now will be Berry Exclusive.

Technically Burmy and Cherrubi can also be gotten from the farm, but not sure if we want a lot of exclusives in the farm, that might be annoying.

* Fix Maranga Mutation

* Refactor flavor evolutions

A bit harder to get now

* Hide locked berries in Berry Master Trades

* Boost Mutation rate for Second Gen

Should make them a bit easier than the first flavor mutations now, which seem to be the first mutations the player gets at the moment.

* Fixing Mulches (again)

* Fixing Mutations and also fixing Berry Quests

Some mutations were crashing. Thankfully no one's reached them yet.

Also fixed quests to only have unlocked berries and also balanced reward.

* Implement Enigma Mutation Hint

* Change to region locked instead

* Revert "Remove Combee from Shop"

This reverts commit d51649cd24fe69dbf682734c3a7c74d10459c66c.

* Add starters to wandering pokemon

* Fix Wepear stage time

* Bump Berry ID by 1 so Cheri gets #01

* Fix for Berry Statistics Modal

* Fix updating Kanto  Berry Master hint every day

* Add mulch description tooltips on mulch list

* Remove tooltips when not visible

Seems to fix some of the lag issues.

* Simplify fix

Now the clicks will work correctly since it's still on the same element. Now the plots are just removed entirely when the farmView isn't visible.

* Revert "Simplify fix"

This reverts commit 3e545074aee3d539b5a680acc28dda0b514bd81e.

* Revert "Remove tooltips when not visible"

This reverts commit 2cd98436134fd72dfa13b8d132376fa04a235136.

* Remove show function

This was causing the lag after hovering over tooltips too many times. The tooltips weren't being handled properly afterwards, which caused popper.js to attempt to reposition them when scrolling, which caused the lag.

* Update src/scripts/farming/Farming.ts

Co-authored-by: Danial <RedSparr0w@users.noreply.github.com>

* Update src/scripts/farming/Farming.ts

Co-authored-by: Danial <RedSparr0w@users.noreply.github.com>

* Update src/scripts/quests/QuestHelper.ts

Co-authored-by: Danial <RedSparr0w@users.noreply.github.com>

* Update src/scripts/quests/QuestHelper.ts

Co-authored-by: Danial <RedSparr0w@users.noreply.github.com>

* Update src/scripts/quests/QuestHelper.ts

Co-authored-by: Danial <RedSparr0w@users.noreply.github.com>

* Fix tabs

* Add more statistics

* Added shovel statistics

* Fix flavor mutations to be possible

* Remove +

Co-authored-by: Danial <RedSparr0w@users.noreply.github.com>