change(dungeon): Add Solaceon Ruins (#1209)
* Add Solaceon Ruins

Added Solaceon Ruins with 4 unown types (solacen - discor = laen), 2 of them are brand spanking new. Also moved Hippopotas to these ruins. Reasoning for that is Hippopotas is currently on 214 because that's where the Maniac Tunnel is, which is the only place that has wild Hippopotas in all of Sinnoh. The tunnel does connect to some secret rooms in the Solaceon Ruins, so while the Maniac Tunnel is still not in the game (probably won't happen either), I think having Hippopotas in a dungeon makes sense, and having them in the Solaceon Ruins also makes sense.

* Update GameConstants.ts

* Update Update.ts

* MapStuff

* Update Dungeon.ts

Co-authored-by: Danial <RedSparr0w@users.noreply.github.com>