change(dungeon): Only allow to move to tiles next to visited (#1547)
* Only allow players to move to tiles adjacent to tiles that have been visited

Add chest image to chest tiles.
Add boss image to boss tiles.

* Fixup if amount is missing

Not sure why this would happen, but it happens 🤷‍♂️

* Return MissingNo. without HP if invalid route

Fixes Rattata appearing if you were on a dungeon when saving then reloading.
Shouldn't actually appear ever, but a decent failsafe as users are more likely to report a MissingNo. rather than a Rattata.