change(html): Credits modal added to start menu (#1553)
* Added some missing trainer sprites and Alola backgrounds

There are two backgrounds for the dungeon "Nebby" (listed in Dungeon.ts as of 1st Sept 2021) as the corresponding location in the mainline games is ambiguous. The image titled "Altar of the Sunne and Moone" refers to the town / gym.

* Fixup name

* Piggy backing

Piggy backing off this PR to reduce some other trainer image sizes

* Credits modal added to start menu

A place to credit sprite contributors. Beautifying it in HTML is beyond me at the moment, so it'll stay plain for now.

* Update IanWalder link/name

Co-authored-by: IanWalder <76807453+IanWalder@users.noreply.github.com>

* minor fixup

* Add CodeCredits, and some more SpriteCredits, Add images

Co-authored-by: Danial <admin@redsparr0w.com>
Co-authored-by: Danial <RedSparr0w@users.noreply.github.com>