Getting rid of dungeon level
No more dungeon level, I also changed the line return at the end of the first dungeon so that it's the same in the whole file.
Everything seems to work fine. Only question is: should I create a function to get a dungeon level as it's already the same call in three places ?