More held items
All that's really left now is Dragon Fang. Victory Road is still open but that's kinda lame so maybe it's gonna be a shard thing instead or something like that IDK.

There's also still Hard Stone and Metal Coat which are an underground item and an evolution item so that's a whole thing.

Anyway, other than that all the type based items are now taken care of.