Updated Mind and Pixie Plate sprites
Fixes #3372
Currently, the sprites for Mind and Pixie Plates are a mix of newer sprites for the icons and older sprites for how they appear in the underground, leading to colors used for both being inconsistent with each other and almost appearing to be swapped.

This adjusts the colors of the underground sprites to match both the icons used in Pokeclicker and how the plates appear in BDSP's Grand Underground.

The "downside" is that the Mind Plate is now less accurate to how it appeared in Gen 4's underground.