fix: bug in underground item placement
Currently, if y is 10 and reward.space.length is 2 (i.e. we're on the second to last row and the item is two spaces high), then the item is placeable, but current comparison is '>= 12' so the method returns false. Using strict comparison fixes this issue.

Additionally, Mine.getRandomCoord also never returns the last row/column, adding one to the call fixes that.