Catch or Hatch Patch.
Adjusts all quest-text where Hatching is a valid strategy to mention it specifically. This patch at this point does not adjust any existing Capture or Hatch quests function, it just includes the hatch notifier if it's even slightly possible. Even cases where hatching is only theoretically possible with a lucky roll of the Shiny Code.