fix(text): Catch or hatch patch (#3863)
* Manaphy Egg Quest

A manaphy quest based on the three Pokemon Ranger Questlines.

* Tabbing errors

Removed and fixed Tabs

* Improved Quest Text

Improved quest text to be more readable in terms of directions on location, to be better read.
Also updated placeholder image of Strange Boulders and the Go-Rock commanders to their proper versions, and properly credited Soul and Drakay. for their hard work.

* Revert "Improved Quest Text"

This reverts commit b61fd992954dabc8645a27b304c8d44a4746a0e7.

* Revert "Revert "Improved Quest Text""

This reverts commit a1a929825bf0081ba66b949c89af18f5c1b82519.

* Typos and Temp Battle Rebalancing

Fixed Typos and buffed the temp battles in the Manaphy Quest.

* .10.8 realignment

Moves update materials to 10.8 instead of 10.7

* 10.9 Updort

* Fixing package Conflicts

* .10.9 Realignment Again

* Text fixes per feedback

* Second Typo Fix

* Update Update.ts

* Update Update.ts

* Update Update.ts

* Update Update.ts

* Typos and Dungeon Boss Cleanup

* Update src/scripts/towns/TownList.ts

Per Aegyo/Jessica's suggestion

Co-authored-by: Jessica Chen <jessica.scoltock@protonmail.com>

* Update src/scripts/towns/TownList.ts

Per Aegyo/Jessica's suggestion

Co-authored-by: Jessica Chen <jessica.scoltock@protonmail.com>

* Update QuestLineHelper.ts

* Revert "Update QuestLineHelper.ts"

This reverts commit da170669ff064e2640aa7a4000936c23ecd05a43.

* Catch or Hatch Patch.

Adjusts all quest-text where Hatching is a valid strategy to mention it specifically. This patch at this point does not adjust any existing Capture or Hatch quests function, it just includes the hatch notifier if it's even slightly possible. Even cases where hatching is only theoretically possible with a lucky roll of the Shiny Code.

* Adjustments to text per feedback.

* CaptureSpecificPokemonQuest fix.

* eslint-fix

* and -> or

* Code cleanup

* Fix?

---------

Co-authored-by: Jaaslet <jonas_bms@hotmail.com>
Co-authored-by: Jessica Chen <jessica.scoltock@protonmail.com>