Swapped access order for Ecruteak gym access and Legendary Beasts
To stay true to the main games, it is now required to pass route 37 to start the Legendary Beasts questline and requires beating Silver inside said questline in order to challenge Morty's gym. The only thing I'm unsure of, and hopefully whoever checks this knows more than me, is if the quest should no longer be a bulletin board quest if it's required for progression. If so, I can attempt to change it.