Updated to Lucho's suggestion
Instead of having Morty's gym and the quest interact in any way, Lucho suggested simply making the legendary beasts quest active when route 37 is cleared as a bulletin which imo makes plenty of sense. This way, the beasts are still available pre-Morty but do not become a required quest that some people may not want to clear.