change(johto): Legendary Beasts quest can be accessed after route 37 (#3975)
* Swapped access order for Ecruteak gym access and Legendary Beasts

To stay true to the main games, it is now required to pass route 37 to start the Legendary Beasts questline and requires beating Silver inside said questline in order to challenge Morty's gym. The only thing I'm unsure of, and hopefully whoever checks this knows more than me, is if the quest should no longer be a bulletin board quest if it's required for progression. If so, I can attempt to change it.

* Update QuestLineHelper.ts

* Updated for no Bulletin Board

Swapped the req to start the quest from Bulletin Board to Sudowoodo (I couldn't get it to activate on route 37 clear for whatever reason)

* Updated to Lucho's suggestion

Instead of having Morty's gym and the quest interact in any way, Lucho suggested simply making the legendary beasts quest active when route 37 is cleared as a bulletin which imo makes plenty of sense. This way, the beasts are still available pre-Morty but do not become a required quest that some people may not want to clear.

---------

Co-authored-by: Jaaslet <jonas_bms@hotmail.com>