Changed the text in Easter Quest Line
Made the text of the Surprise Togepi Quest Line a bit better. Removed the ferry part about Johto and mentioned the forest, since people were having issues with that step. Also made the third one less explicit since it was the only one that named the dungeon.

I still feel this quest needs to change further but like this it's, at least, a bit more clearer for the player where they need to go in the Johto step.