Necrozma, Cosmog, Cosmoem, Solgaleo, Lunala
Added Golbat to Mount Lanakila to keep it at a higher boss count than Lake of the Sunne and Moone, and put it on the same level as Kyurem.
Added Solrock and Lunatone as bosses in Lake of the Sunne and Moone before Cosmog is captured, and Cosmoem after it is captured. Changed the chance of Light items accordingly
Removed Forretress from P2 Lab as non canon