Weather changes
Renamed overcast to cloudy. Updated color to reflect most current colors based on bulbapedia weather page.

Not changing sunny to harsh sunlight because sunny is easier for everyone.

Updated weather boost and removed weather nerfs because damage nerfs are not fun so that we don't have to go with canon for that.

Changed weight of weathers to be uniform. Removed Hail and windy.

Snowy form castform is already triggered by snow and blizzard so removal of hail does not effect the pokemon.
Cherrim (overcast) is just the name of the pokemon and overcast>cloudy change does not effect it.