change(dungeons): Rebalancing Number of bosses in Legendary dungeons (#4160)
* Rayquaza

Rayquaza was increased from 2 bosses to 4 to balance it in line with Celebi. encounters to replace the new bosses are from ORAS

* Azelf, Uxie and Giratina

Saturn, Jupiter and Cyrus have their weights increased to 2 to make Azelf, Uxie and Giratina 1 in 3 (arbitrary minimum maintained from Johto onwards). Grisious orb chance is balanced accordingly (with a slight boost)

* Darkrai

Adds 2 bosses to Newmoon Island.
Fixes error in Brelooms HP in Flower Paradise. Also added Lunatone to Fullmoon Island to match Newmoon Island

* Regigigas

Adds Steelix from DP as a boss to Snowpoint Temple

* Heatran and Volcarona

Moves Camerupt to be a boss in Stark Mountain.
Increases Perry's weight to 2 (Volcarona is the only non legendary pokemon that will be affected by this PR, due to being treated as one in some cases)

* Terrakion

Decreased Victory Road Unova down to 3 bosses so Terrakion matches Coballion and Virizion

* Reshiram and Zekrom

Increased Dragonspiral Tower to 4 bosses, changed DNA chance accordingly

* Keldeo (Resolute)

Added 2 bosses to Pledge Grove

* Genesect

I have no idea what went wrong with this dungeon. almost all the encounters were non canon despite having plenty canon encounters. Also decreased the bosses to 5. Still quite high, but not utterly ridiculous as it used to be

* Diancie and Volcanion

Increased Emil, Monique and Aliana's weights to give the dungeons a total weight of 4 (the arbitrary minimum I am using for most (if not all) legendaries from Kalos onward)

* Necrozma, Cosmog, Cosmoem, Solgaleo, Lunala

Added Golbat to Mount Lanakila to keep it at a higher boss count than Lake of the Sunne and Moone, and put it on the same level as Kyurem.
Added Solrock and Lunatone as bosses in Lake of the Sunne and Moone before Cosmog is captured, and Cosmoem after it is captured. Changed the chance of Light items accordingly
Removed Forretress from P2 Lab as non canon

* Tapu Koko, Tapu Lele, Tapu Bulu, Tapu Fini

Added pokemon that match the tapus types to their boss list. Removed these pokemon from the regular list  (and Gloom, to keep the number matching)

* Guzzlord

Added Zygarde to Resolution Cave

* Change Cosmog fodder to Minior

* Xerneas and Yveltal

updated to be slightly rarer due to their hightened catch rates

---------

Co-authored-by: Jaaslet <jonas_bms@hotmail.com>