Added Notification to auto-proc quests
Added a notification to all quests that auto-proc (except the tutorial). With this, the player can notice when a new Quest Line gets added to their quest list, be it due to completing a gym, a temporary battle, or whatever other event that causes Quest Lines to start.

Also, modified bulletinBoardModal to use the default, as in, removed the parameters it had since they were unnecessary.