Add back non-canon areas and NPC rocks
Added Moss and Ice rocks as NPCs to all Leafeon and Glaceon evolution areas from the games. They appear in Hoenn so they have been given a Sinnoh requirement.
Added back old dungeons to the environments for returning players, and because the Moss and Ice rocks mark evolution areas but can only be put in towns as NPCs. It is easier to extend an environment to a dungeon than add interactable Rocks to routes.
Added Alola Route 12 to Magnetic Field environments because we don't have Blush Mountain, and it will help players come across Vikavolt more easily until we add the area (preferably with an NPC).
Changed Snowpoint NPC to say Route 217 so players can learn that environment evolutions aren't solely dungeon-dependent.