Added Shadow Modifier to the FAQ
Added Shadow Modifier to the FAQ in the EVs section since it was missing. Made Base Wanderer Yield say 0EV instead of 0xEV since it looked wrong when compared to the other values.
Finally, put values with the same Yield / Modifier on the same line to make the list take less space.