change(requirements): Fixing the unlocked routes in Johto (#4838)
In Johto, before defeating Morty around Ecruteak, Mt. Mortar is unlocked, but Route 42, Mahogany Town, and Route 43 are not. In HGSS, it is possible to access Route 42 (namely it's grass for encounters), as well as everywhere else mentioned, in 2 ways. Either you can beat Morty and obtain Surf, or you can go through Mt. Mortar prior to beating Morty. You are not able to get to the Red Gyarados without Surf, so you still cannot start that questline.

Co-authored-by: Qwertypop04 <63805905+Qwertypop04@users.noreply.github.com>