Changed Gym Quest Logic to Only Generate For Cleared Gyms
Currently, Defeat Gym Quests was looking at which gyms were unlocked to the player. This caused situations where the game would ask, for example, for the player to beat a Champion as soon as they were done with the Elite 4 thus making it an "impossible" quest for a bit. This changes the logic from isUnlocked() to clears() ensuring it only picks gyms the player has already cleared at least once.
Also, makes Defeat Gym Quests only available after obtaining the first badge (Boulder).