Burmy and Caves
Burmy limited to canon because I didn't want to spend time nor set the expectation to put every single instance where it might change forms. Note: Canon was used as a guideline and Burmy bears no strict adherence to it because the game version depends on the encounter tile as opposed to an entire rotue.
Added Cave environment for easier Burmy forms and whatever else they might be used for in the future.