Changed Eternapantatus image for the correct Eternapants
The Lair of Giants quest line was using regular Eternapantus instead of the Eternapants version. This PR fixes it so it uses the right image as the text mentions the Pokémon fleeing to the south but the one that does is the Eternapants version and not regular one that becomes unlocked in Energy Plant.