requirements(dungeon): Dungeons can have their own requirements (#5000)
Dungeons can now have separate requirements from their towns. This lets us use dungeon towns without unlocking dungeons. This can be used for story towns like the Plasma Frigates or Slumbering Weald, where it is preferred to have npcs or temporary battles in the dungeon town be accessible, but not the dungeon itself.

---------

Co-authored-by: Qwertypop04 <63805905+Qwertypop04@users.noreply.github.com>