feature(helditems): Everstone (#5185)
* Everstone

Hello ! I've added the item "Everstone_Held" to prevent the evolutions by leveling, the code works perfectly. The item can be hold only if the Pokémon has evolution by leveling and the player haven't the evolution.

However, there are three issues:
1) The Everstone and Everstone_Held items are two different items. A solution may be another diplayname for Everstone_Held which is "Everstone".
2) I allow the use of Everstone_Held at Kanto. I know the held items are in Johto but I find this fun to have this item quickly.
3) At this point, the item can be bought at Viridian City Shop for 100 Pokédollars. It's a placeholder for the tests. I let you decided the best way to obtain the item.

So !

I'm happy to participate to make Pokéclicker better.

Aenonis

* npm run eslint fixes

* enlarge to all pokemons

enlarge the Everstone_Held to all pokémons (i.e. stone evolutions). Issue : The gray pokeball icon doesn't display "Pokémon hold Everstone" and is not blocked. However, evolution is blocked if I press the evolve button on ItemEvolution, as expected.

* evo.trigger none fixes

I've forgot to exclude the evolutions who have no trigger.

* rework of the code

Rework the code to use the classes Restrictions. Remove the code on PartyPokémon classe and dispatch in the adequates files. Thanks to TimeSpazm to help me !

* limit to available pokémons

don't use the everstone on the pokémon who have evolutions on regions not already reached

* according price

according price to TypeRestrictedAttackBonusHeldItem even if it can't be bought

* mégas

add the management of the megas

* Babies

add the babies implementation

* npm run eslint-fix

* Code clean up, improvements

---------

Co-authored-by: CypherX <cypherx00@gmail.com>