change(battlefrontier): Added Random Environment to Battle Frontier (#5195)
* Made Charity Chansey line be only female

Fixes the most worstest bug of all time.

* Update translations

* Made BF Change Environment Every 25 Stages

Battle Frontier is boring. Grass is boring. So now more than Grass can happen. This does trigger evolutions which may or may not be a concern.

* No Env Evo in Battle Frontier

* Simplified code

* translation stuff

* Some extra changes

Made Environment be Default when starting a new run.
Made it so the current Environment being used couldn't be selected at the 25 switch so it doesn't happen two times in a row.

* preserve environment when resuming

doesn't make sense to reset it. I know a reload resets it but that's an entirely different case

---------

Co-authored-by: Jaaslet <jonas_bms@hotmail.com>
Co-authored-by: CypherX <cypherx00@gmail.com>