change(map): Swapped Underground with Farm (#5238)
Swapped the locations of the Underground and the Farm. With this, I also changed the requirements making the Farm unlock as soon as Route 6 is cleared (or as soon as the player can access Route 11).
I removed Farming's "has route access" check since the KeyItem already has this check and I found it unnecessary. Unless it's linked to something but I can't see it.

Co-authored-by: Qwertypop04 <63805905+Qwertypop04@users.noreply.github.com>