Some Small Fixes
Changed Base Wanderer Yield in the FAQ from 0 to 0.1 and removed the line about Base Wanderers not receiving EVs as that no longer the case.
Reversed Category sort so it maintains the same order from live.
Removed Pinkan Pikachu reward from the Temporary Battle as now it's meant to be acquired as a mimic.