refactor(environments): Environment refactor (#4610)
* Add SubEnvironment base code

* Add more SubEnvironment code

For Gyms and Temp Battles

* Update calculateBattleCssClass

uses SubEnvironment by default if there is one

* Add SubEnvironment Record

* Cave Environment clean up

* Indoor Environment clean up

* Add comments

Burmy :3

* Remove leftover constants

* Alola cemeteries aren't Indoors

* eslint

* Use only SubEnvironments for css and Environments for game mechanics

* Change Environments to game mechanic only

Environments are now only for game mechanics (evolutions and dive ball)
Changed evolutions in PokemonList but kept Dummy ones for egg steps; to be removed later

* names

* use EnvironmentData for both Records

* Simplify environments only for evolution uses

* Gyms and TempBattles

* Add Water Cave

Both a cave AND water!

* canon i guess

* Update hint to have space

* condense evolutions

* Update PokemonList.ts

* fix mistakes

* shorten comment

* hisui

* Add back non-canon areas and NPC rocks

Added Moss and Ice rocks as NPCs to all Leafeon and Glaceon evolution areas from the games. They appear in Hoenn so they have been given a Sinnoh requirement.
Added back old dungeons to the environments for returning players, and because the Moss and Ice rocks mark evolution areas but can only be put in towns as NPCs. It is easier to extend an environment to a dungeon than add interactable Rocks to routes.
Added Alola Route 12 to Magnetic Field environments because we don't have Blush Mountain, and it will help players come across Vikavolt more easily until we add the area (preferably with an NPC).
Changed Snowpoint NPC to say Route 217 so players can learn that environment evolutions aren't solely dungeon-dependent.

* Update GameConstants.d.ts

* Default env to Outdoors

There's no evolution method for Burmy (Plant) yet, but just in case we do form changes in the future

* Merge correction

* curly brace...

* Add Cactus Desert

* Create InSubEnvironmentRequirement.ts

* Update InSubEnvironmentRequirement.ts

* Update globals.ts

* Update temporaryWindowInjection.ts

* Remove requirement, no need for it

* burmies

* multienvironment support

Changes getCurrentEnvironment to return an array
This lets us assign multiple environments to the same area

* merge cave environments

* Water environment for Diveball

* correct hisui burmy

* Outdoors = not Cave or Indoors

* Pluralize

getCurrentEnvironment can return multiple envs so I pluralized it and env to make more sense

* Comment

* gym and temp battles

* Water touch ups

fixed Comments
changed Water routes map css back

* rename SubEnv -> BattleBackground

* eslint semicolon

* streamline hisui envs

* Wormadam doesn't change forms

* eslint

* Update TownList.ts

* Remove added battle backgrounds

These will be added in another PR to keep this one more focused and easier to understand

* Update GameConstants.ts

* Update GymList.ts

* stuff

- environments are now arrays
- changed burmy environment names
- expanded getCurrentEnvironments to streamline some areas

* Update GymList.ts

* don't need these

* Update GameConstants.ts

* Update GymList.ts

* Update MapHelper.ts

* eslint

* Burmy and Caves

Burmy limited to canon because I didn't want to spend time nor set the expectation to put every single instance where it might change forms. Note: Canon was used as a guideline and Burmy bears no strict adherence to it because the game version depends on the encounter tile as opposed to an entire rotue.
Added Cave environment for easier Burmy forms and whatever else they might be used for in the future.

* coronet stone

* generalize function for wiki use

* merging

* Valencia and Pinkan town into route

These were named instead of numbered and they weren't working

---------

Co-authored-by: Qwertypop04 <63805905+Qwertypop04@users.noreply.github.com>