Regional Debuff
Changed Regional Debuff from a global value that goes from 0.2 maintained to 1, to comparing a Pokémon's native region vs the player's current region and reducing from 1 (native) to 0.25 minimum. It goes in decrements of 0.1 and it starts at 0.45 (1 region difference) and goes down to 0.25 (3 regions difference) where it reaches its limit.