chore(deps): Add support for recursively importing html files (#5435)
* Replace deprecated gulp-html-import package

Replaced deprecated package with a local version modified to avoid vulnerabilities, deprecated behavior, and unnecessary packages.

* Upgrade HTML importing to support nested imports

* Move townMap out of index.html to an import

PrDemonstration that the recursive importing both works and is immediately useful. The town map was previously stuck in index.html because it needed to import all the region SVGs. Now it can be a separate imported file without causing any issues.

* Tweak error message

* Don't update vitest to breaking version

* Require that @import statements get their own line

* Workaround for npm bug...?

* fixup linting warnings

---------

Co-authored-by: CypherX <cypherx00@gmail.com>
Co-authored-by: CypherX <672420+CypherX@users.noreply.github.com>