Added the Breeding Queue to the Breeding Modal
The Breeding Queue is now visible inside the Modal. In addition, an option is available to disable this in case the player wants to stick with how the modal used to look. The file breedingModal INLINETEST.html is an alternative take on what the queue could look like in the modal.